got the Trade Manager to work.
2/3 managers are completed. the upgrade manager still needs the buttons
to be implemented. The save functionality needs to be done, and then the
settings need to actually do something as well. From what I gather,
events will be near impossible to serialize in their current form. I'm
going to need to add an extra required member variable to save the type
of the event and find out how actionListeners are serialized