Introduced game assets and board is now working
Player pieces must be added in and the dice could use some work. Maybe
even remove them from the board and have them in their own class
altogether so that they can be displayed in a separate panel next to the
Message Panel. 
Next work to be done on the Message Panel to create the needed panel,
the controller classes or "Messages" and how much knowledge to give to
each object