I believe it is safe to say that Monopoly is now complete!
The main exceptions are the instruction manuals, tests, and proper
documentation but for the casual player, nothing is missing. Bugs will
inevitably occur. From now on, I will make an effort to include the jar
in the main folder of the file so only a repository pull is needed to
update everything