Opening changes to custom textures and changed logger API
Massive changes happened with stickers, stamps, and GameTokens. Totally
reinvented the wheel on the first one, overhauled the second, and
generalized the third for other board game applications

Switched from a personally made logger (called LogMate) to Google
Flogger.

Currently trying to fix a bug where, when starting a game with a custom
texture, all icons are forced to the top left corner and cannot move
from there